requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiohttp-cors==0.7.0
aiosignal==1.3.1
albumentations==1.4.7
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.2.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
astroid==3.0.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Authlib==1.3.0
autopep8==2.0.4
autotrain-advanced==0.7.118
Babel==2.14.0
beautifulsoup4==4.12.2
black==24.8.0
bleach==6.1.0
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.3
cattrs==24.1.2
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
chardet2==2.0.3
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cloup==3.0.5
codecarbon==2.3.5
colorama==0.4.6
colorful==0.5.6
colorlog==6.8.2
comm==0.2.1
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
Cython==3.0.10
datasets==2.19.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.27.2
dill==0.3.8
distlib==0.3.8
Django==5.0.1
django-filter==23.5
djangorestframework==3.14.0
dm-tree==0.1.8
dnspython==2.6.1
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.7.0
email_validator==2.1.1
evaluate==0.4.1
exchange_calendars==4.5.6
executing==2.0.1
fairscale==0.4.13
Farama-Notifications==0.0.4
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.19.1
filelock==3.13.1
fire==0.5.0
flake8==7.1.1
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
fonttools==4.47.2
fqdn==1.5.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.2.0
gguf==0.6.0
gitdb==4.0.11
GitPython==3.1.43
glcontext==2.5.0
google-api-core==2.19.0
google-auth==2.30.0
googleapis-common-protos==1.63.1
grpcio==1.64.0
gymnasium==0.28.1
h11==0.14.0
harperdb==0.4.2
hf_transfer==0.1.6
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.25.2
identify==2.6.0
idna==3.6
imageio==2.34.1
importlib_metadata==7.1.0
inflate64==1.0.0
iniconfig==2.0.0
invisible-watermark==0.2.0
ipadic==1.0.0
ipykernel==6.28.0
ipython==8.20.0
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.13.2
isosurfaces==0.1.2
itsdangerous==2.2.0
jax-jumpy==1.0.0
jedi==0.19.1
Jinja2==3.1.3
jiwer==3.0.3
joblib==1.4.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.4
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kaggle==1.6.14
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
lazy-object-proxy==1.10.0
lazy_loader==0.4
lightning-utilities==0.11.2
linkify-it-py==2.0.3
-e git+https://github.com/meta-llama/llama.git@be327c427cc5e89cc1d3ab3d3fec4484df771245#egg=llama
loguru==0.7.2
lxml==5.3.0
lz4==4.3.3
Mako==1.3.5
manim==0.18.1
ManimPango==0.5.0
mapbox-earcut==1.0.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdit-py-plugins==0.4.1
mdurl==0.1.2
memray==1.13.0
mistune==3.0.2
moderngl==5.10.0
moderngl-window==2.4.6
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multipledispatch==1.0.0
multiprocess==0.70.16
multitasking==0.0.11
multivolumefile==0.2.3
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.14.1
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nodeenv==1.9.1
nose==1.3.7
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.0
nvidia-ml-py==12.535.161
nvitop==1.3.2
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
optuna==3.6.1
orjson==3.10.3
overrides==7.4.0
packaging==24.0
pandas==2.2.2
pandas_market_calendars==4.4.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.12.1
peewee==3.17.6
peft==0.11.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.1.0
pluggy==1.5.0
pre-commit==3.8.0
pretty-errors==1.2.25
prometheus-client==0.19.0
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==4.23.4
psutil==5.9.7
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py-spy==0.3.14
py7zr==0.21.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybcj==1.0.2
pycairo==1.26.1
pycocotools==2.0.7
pycodestyle==2.12.1
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
pyflakes==3.2.0
pyglet==2.0.15
Pygments==2.17.2
pylint==3.0.3
pyluach==2.2.0
pyngrok==7.1.6
pynvml==11.5.0
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyparsing==3.1.1
pyppmd==1.1.0
pyrate-limiter==2.10.0
pyrr==0.10.3
pytest==8.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2023.3.post1
PyWavelets==1.6.0
PyYAML==6.0.1
pyzmq==25.1.2
pyzstd==0.16.0
qtconsole==5.5.1
QtPy==2.4.1
rapidfuzz==3.9.3
ray==2.30.0
referencing==0.32.1
regex==2024.5.15
requests==2.31.0
requests-cache==1.2.1
requests-ratelimiter==0.7.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rouge_score==0.1.2
rpds-py==0.17.1
rsa==4.9
sacremoses==0.1.1
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.11.4
screeninfo==0.8.1
Send2Trash==1.8.2
sentence-transformers==3.0.0
sentencepiece==0.2.0
sentry-sdk==2.3.1
seqeval==1.2.2
setproctitle==1.3.3
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
skia-pathops==0.8.0.post1
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.4.4
srt==3.5.3
stack-data==0.6.3
starlette==0.37.2
svgelements==1.9.6
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
terminado==0.18.0
text-unidecode==1.3
texttable==1.7.0
textual==0.70.0
threadpoolctl==3.5.0
tifffile==2024.5.22
tiktoken==0.6.0
timm==1.0.3
tinycss2==1.2.1
tk==0.1.0
tokenizers==0.20.1
toml==0.10.2
tomlkit==0.12.3
toolz==1.0.0
torchmetrics==1.4.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
transformers==4.45.2
trl==0.8.6
typer==0.12.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tyro==0.8.4
tzdata==2023.4
uc-micro-py==1.0.3
ujson==5.10.0
uri-template==1.3.0
url-normalize==1.4.3
urllib3==2.1.0
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.26.2
wandb==0.17.0
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.2
widgetsnbextension==4.0.9
wrapt==1.16.0
xgboost==2.0.3
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.1
